+++ src/FileEmbedLzma.hs
@@ -48,11 +48,9 @@ import qualified Data.Text               as T
 import qualified Data.Text.Lazy          as LT
 import qualified Data.Text.Lazy.Encoding as LTE
 
-#if MIN_VERSION_template_haskell(2,16,0)
 import qualified Data.ByteString.Internal as BS.Internal
 
-import Language.Haskell.TH.Syntax (Bytes (..))
-#endif
+import Language.Haskell.TH.Syntax (Bytes (..), liftTyped)
 
 -- $setup
 -- >>> :set -XTemplateHaskell -dppr-cols=9999
@@ -88,18 +86,13 @@ makeAllRelative topdir = map (first (("/" ++) . makeRelative topdir))
 
 -- | Makes lazy 'LBS.ByteString' expression.
 -- Embedded value is compressed with LZMA.
-lazyBytestringE :: LBS.ByteString -> Q Exp
+lazyBytestringE :: LBS.ByteString -> Code Q LBS.ByteString
 lazyBytestringE lbs =
-    [| LZMA.decompress . LBS.fromStrict . unsafePerformIO |] `appE`
-    ([| BS.Unsafe.unsafePackAddressLen |] `appE` l `appE` s)
+    [|| LZMA.decompress (LBS.fromStrict (unsafePerformIO (BS.Unsafe.unsafePackAddressLen $$(liftTyped l) $$(unsafeCodeCoerce s)))) ||]
   where
     bs = LBS.toStrict $ LZMA.compressWith params lbs
-#if MIN_VERSION_template_haskell(2,16,0)
     s = litE $ bytesPrimL $ bsToBytes bs
-#else
-    s = litE $ stringPrimL $ BS.unpack bs
-#endif
-    l = litE $ integerL $ fromIntegral $ BS.length bs
+    l = BS.length bs
 
     params = LZMA.defaultCompressParams
         {- doesn't seem to affect much
@@ -108,15 +101,12 @@ lazyBytestringE lbs =
         }
         -}
 
-#if MIN_VERSION_template_haskell(2,16,0)
 bsToBytes :: BS.ByteString -> Bytes
 bsToBytes (BS.Internal.PS fptr off len) = Bytes fptr (fromIntegral off) (fromIntegral len)
-#endif
 
-makeEmbeddedEntry :: Name -> (FilePath, (Int64, Int64)) -> Q Exp
-makeEmbeddedEntry name (path, (off, len)) = do
-    let y = [| LBS.toStrict . LBS.take len . LBS.drop off |] `appE` varE name
-    [| (,) path |] `appE` y
+makeEmbeddedEntry :: Code Q LBS.ByteString -> (FilePath, (Int64, Int64)) -> Code Q (FilePath, BS.ByteString)
+makeEmbeddedEntry name (path, (off, len)) =
+    [|| (path, LBS.toStrict (LBS.take len (LBS.drop off $$name))) ||]
 
 concatEntries :: Traversable t => t LBS.ByteString -> (LBS.ByteString, t (Int64, Int64))
 concatEntries xs = (bslEndo LBS.empty, ys)
@@ -135,24 +125,14 @@ concatEntries xs = (bslEndo LBS.empty, ys)
 -- Directories
 -------------------------------------------------------------------------------
 
+
 -- | Embed a @[('FilePath', 'Data.ByteString.ByteString')]@ list, traversing given directory.
-embedDir :: FilePath -> Q Exp
-embedDir topdir = do
+embedDir :: FilePath -> Code Q [(FilePath, BS.ByteString)]
+embedDir topdir = joinCode $ do
     pairs' <- runIO $ listDirectoryFiles topdir
     for_ pairs' $ qAddDependentFile . fst
     let pairs = makeAllRelative topdir pairs'
-    embedPairs pairs
-
-embedPairs :: [(FilePath, LBS.ByteString)] -> Q Exp
-embedPairs pairs = do
-    -- we do a hop to only embed single big bytestring.
-    -- it's beneficial as lzma have more stuff to compress
-    let (bsl, Compose offsets) = concatEntries (Compose pairs)
-    bslName <- newName "embedBsl"
-    bslExpr <- lazyBytestringE bsl
-    let e = letE [ return $ ValD (VarP bslName) (NormalB bslExpr) [] ] $
-                listE $ map (makeEmbeddedEntry bslName) offsets
-    sigE e [t| [(FilePath, BS.ByteString)] |]
+    return $ embedPairs pairs
 
 -- | Embed a @[('FilePath', 'Data.ByteString.ByteString')]@ list, recursively traversing given directory path.
 --
@@ -165,18 +145,27 @@ embedPairs pairs = do
 -- @
 --
 --
--- >>> $(embedRecursiveDir "example")
+-- >>> $$(embedRecursiveDir "example")
 -- [("/Example.hs","..."),("/example.txt","Hello from the inside.\n")]
 --
--- >>> :t $(embedRecursiveDir "example")
--- $(embedRecursiveDir "example") :: [(FilePath, BS.ByteString)]
+-- >>> :t $$(embedRecursiveDir "example")
+-- $$(embedRecursiveDir "example") :: [(FilePath, BS.ByteString)]
 --
-embedRecursiveDir :: FilePath -> Q Exp
-embedRecursiveDir topdir = do
+embedRecursiveDir :: FilePath -> Code Q [(FilePath, BS.ByteString)]
+embedRecursiveDir topdir = joinCode $ do
     pairs' <- runIO $ listRecursiveDirectoryFiles topdir
     for_ pairs' $ qAddDependentFile . fst
     let pairs = sort (makeAllRelative topdir pairs')
-    embedPairs pairs
+    return $ embedPairs pairs
+
+embedPairs :: [(FilePath, LBS.ByteString)] -> Code Q [(FilePath, BS.ByteString)]
+embedPairs pairs = do
+    -- we do a hop to only embed single big bytestring.
+    -- it's beneficial as lzma have more stuff to compress
+    let (bsl, Compose offsets) = concatEntries (Compose pairs)
+    let e = [|| let embedded = $$(lazyBytestringE bsl)
+                in $$(typedListE (map (makeEmbeddedEntry [|| embedded ||]) offsets)) ||]
+    typedSigE e [t| [(FilePath, BS.ByteString)] |]
 
 -------------------------------------------------------------------------------
 -- Strings
@@ -184,41 +173,51 @@ embedRecursiveDir topdir = do
 
 -- | Embed a lazy 'Data.ByteString.Lazy.ByteString' from a file.
 --
--- >>> :t $(embedLazyByteString "file-embed-lzma.cabal")
--- $(embedLazyByteString "file-embed-lzma.cabal") :: LBS.ByteString
+-- >>> :t $$(embedLazyByteString "file-embed-lzma.cabal")
+-- $$(embedLazyByteString "file-embed-lzma.cabal") :: LBS.ByteString
 --
-embedLazyByteString :: FilePath -> Q Exp
-embedLazyByteString fp = do
+embedLazyByteString :: FilePath -> Code Q LBS.ByteString
+embedLazyByteString fp = joinCode $ do
     qAddDependentFile fp
     bsl <- runIO $ LBS.readFile fp
-    lazyBytestringE bsl
+    return $ lazyBytestringE bsl
 
 -- | Embed a strict 'Data.ByteString.ByteString' from a file.
 --
--- >>> :t $(embedByteString "file-embed-lzma.cabal")
--- $(embedByteString "file-embed-lzma.cabal") :: BS...ByteString
+-- >>> :t $$(embedByteString "file-embed-lzma.cabal")
+-- $$(embedByteString "file-embed-lzma.cabal") :: BS...ByteString
 --
-embedByteString :: FilePath -> Q Exp
-embedByteString fp = [| LBS.toStrict |] `appE` embedLazyByteString fp
+embedByteString :: FilePath -> Code Q BS.ByteString
+embedByteString fp = [|| LBS.toStrict $$(embedLazyByteString fp) ||]
 
 -- | Embed a lazy 'Data.Text.Lazy.Text' from a UTF8-encoded file.
 --
--- >>> :t $(embedLazyText "file-embed-lzma.cabal")
--- $(embedLazyText "file-embed-lzma.cabal") :: LT.Text
+-- >>> :t $$(embedLazyText "file-embed-lzma.cabal")
+-- $$(embedLazyText "file-embed-lzma.cabal") :: LT.Text
 --
-embedLazyText :: FilePath -> Q Exp
-embedLazyText fp = do
+embedLazyText :: FilePath -> Code Q LT.Text
+embedLazyText fp = joinCode $ do
     qAddDependentFile fp
     bsl <- runIO $ LBS.readFile fp
     case LTE.decodeUtf8' bsl of
         Left e  -> reportError (show e)
         Right _ -> return ()
-    [| LTE.decodeUtf8 |] `appE` lazyBytestringE bsl
+    return $ [|| LTE.decodeUtf8 $$(lazyBytestringE bsl) ||]
 
 -- | Embed a strict 'Data.Text.Text' from a UTF8-encoded file.
 --
--- >>> :t $(embedText "file-embed-lzma.cabal")
--- $(embedText "file-embed-lzma.cabal") :: T...Text
+-- >>> :t $$(embedText "file-embed-lzma.cabal")
+-- $$(embedText "file-embed-lzma.cabal") :: T...Text
 --
-embedText :: FilePath -> Q Exp
-embedText fp = [| LT.toStrict |] `appE` embedLazyText fp
+embedText :: FilePath -> Code Q T.Text
+embedText fp = [|| LT.toStrict $$(embedLazyText fp) ||]
+
+-------------------------------------------------------------------------------
+-- TTH helpers
+-------------------------------------------------------------------------------
+
+typedSigE :: Code Q a -> Q Type -> Code Q a
+typedSigE x a = unsafeCodeCoerce $ sigE (unTypeCode x) a
+
+typedListE :: [Code Q a] -> Code Q [a]
+typedListE xs = unsafeCodeCoerce $ listE $ map unTypeCode xs